// src/components/IconButton/types.ts
export enum IconButtonVariant {
    Contained = "contained",
    Outlined = "outlined",
    Ghost = "ghost",
}

export enum IconButtonSize {
    Small = "small",
    Medium = "medium",
    Large = "large",
}

export enum IconButtonRadius {
    Small = "small",
    Medium = "medium",
    Large = "large",
}